Create ast.variable and bug fixes (#637)
Move classes necessary for the handling of Variables into a new file `ast.variable`.
Fixes #490 (ast/core.py too large for codacy).

**Commit Summary**

- Move classes necessary for the handling of Variables into a new file `ast.variable`;
- Move Slice to `ast.internals` (should this be in operators?);
- Remove `is_polymorphic` property from Variables. This property exists in the class dtype and a "polymorphic variable" doesn't make much sense;
- Add `passed_from_dotted` property to `Declare`. In fortran types are declared differently when they are used for arguments which are passed due to class syntax;
- General Codacy fixes.
