+++ pyccel/ast/builtins.py
@@ -16,8 +16,6 @@ from sympy import Expr, Not
 from sympy import sympify
 from sympy.tensor import Indexed, IndexedBase
 
-from pyccel.ast.datatypes import iso_c_binding
-
 from .basic     import Basic, PyccelAstNode
 from .datatypes import (NativeInteger, NativeBool, NativeReal,
                         NativeComplex, NativeString, str_dtype,
