+++ pyccel/ast/mathext.py
@@ -5,8 +5,8 @@
 
 import math
 
+from pyccel.ast.variable  import Constant
 from pyccel.ast.internals import PyccelInternalFunction
-from pyccel.ast.core      import Constant
 from pyccel.ast.datatypes import (NativeInteger, NativeBool, NativeReal,
                                   default_precision)
