+++ pyccel/ast/utilities.py
@@ -14,8 +14,7 @@ import pyccel.decorators as pyccel_decorators
 from pyccel.symbolic import lambdify
 from pyccel.errors.errors import Errors
 
-from .core     import (AsName, Import, FunctionDef, Constant,
-                       Variable, ValuedVariable, FunctionCall)
+from .core     import (AsName, Import, FunctionDef, FunctionCall)
 
 from .builtins      import builtin_functions_dict, PythonMap
 from .itertoolsext  import Product
@@ -24,6 +23,7 @@ from .literals      import LiteralString
 
 from .numpyext      import (numpy_functions, numpy_linalg_functions,
                             numpy_random_functions, numpy_constants)
+from .variable      import (Constant, Variable, ValuedVariable)
 
 __all__ = (
     'build_types_decorator',
