+++ pyccel/codegen/printing/ccode.py
@@ -11,13 +11,12 @@ import operator
 from sympy.core           import Tuple
 from pyccel.ast.builtins  import PythonRange, PythonFloat, PythonComplex
 
-from pyccel.ast.core      import Declare, Slice, ValuedVariable
+from pyccel.ast.core      import Declare
 from pyccel.ast.core      import FuncAddressDeclare, FunctionCall
 from pyccel.ast.core      import Deallocate
-from pyccel.ast.core      import FunctionAddress, PyccelArraySize
-from pyccel.ast.core      import Assign, datatype, Variable, Import
-from pyccel.ast.core      import SeparatorComment, VariableAddress
-from pyccel.ast.core      import DottedName
+from pyccel.ast.core      import FunctionAddress
+from pyccel.ast.core      import Assign, datatype, Import
+from pyccel.ast.core      import SeparatorComment
 from pyccel.ast.core      import create_incremented_string
 
 from pyccel.ast.operators import PyccelAdd, PyccelMul, PyccelMinus, PyccelLt, PyccelGt
@@ -27,12 +26,18 @@ from pyccel.ast.operators import PyccelUnarySub, IfTernaryOperator
 from pyccel.ast.datatypes import default_precision, str_dtype
 from pyccel.ast.datatypes import NativeInteger, NativeBool, NativeComplex, NativeReal, NativeTuple
 
+from pyccel.ast.internals import Slice
+
 from pyccel.ast.literals  import LiteralTrue, LiteralImaginaryUnit, LiteralFloat
 from pyccel.ast.literals  import LiteralString, LiteralInteger, Literal
 from pyccel.ast.literals  import Nil
 
 from pyccel.ast.numpyext import NumpyFull, NumpyArray
 from pyccel.ast.numpyext import NumpyReal, NumpyImag, NumpyFloat
+
+from pyccel.ast.variable import ValuedVariable
+from pyccel.ast.variable import PyccelArraySize, Variable, VariableAddress
+from pyccel.ast.variable import DottedName
 
 
 from pyccel.codegen.printing.codeprinter import CodePrinter
