+++ pyccel/codegen/printing/cwrappercode.py
@@ -16,10 +16,10 @@ from pyccel.ast.literals  import Nil
 
 from pyccel.ast.builtins import PythonPrint
 
-from pyccel.ast.core import Variable, ValuedVariable, Assign, AliasAssign, FunctionDef, FunctionAddress
+from pyccel.ast.core import Assign, AliasAssign, FunctionDef, FunctionAddress
 from pyccel.ast.core import If, Return, FunctionCall
 from pyccel.ast.core import create_incremented_string, SeparatorComment
-from pyccel.ast.core import VariableAddress, Import
+from pyccel.ast.core import Import
 from pyccel.ast.core import AugAssign
 
 from pyccel.ast.operators import PyccelEq, PyccelNot, PyccelAnd, PyccelNe, PyccelOr, PyccelAssociativeParenthesis, IfTernaryOperator
@@ -38,6 +38,8 @@ from pyccel.ast.cwrapper import numpy_check_flag, numpy_flag_c_contig, numpy_fla
 from pyccel.ast.cwrapper import PyArray_CheckScalar, PyArray_ScalarAsCtype
 
 from pyccel.ast.bind_c   import as_static_function_call
+
+from pyccel.ast.variable  import VariableAddress, Variable, ValuedVariable
 
 from pyccel.errors.errors import Errors
 from pyccel.errors.messages import PYCCEL_RESTRICTION_TODO
